MCR-1209 #resolve IIIF-Image iview impl now supports floating point numbers
git-svn-id: https://server.mycore.de/svn/mycore/trunk@34736 7089d063-bc14-496d-ac90-53de1872eb6c

+++ mycore-iview2/src/main/java/org/mycore/iiif/model/MCRIIIFImageTargetRotation.java
@@ -4,9 +4,9 @@ package org.mycore.iiif.model;
 public class MCRIIIFImageTargetRotation {
 
     private final boolean mirrored;
-    private final int degrees;
+    private final double degrees;
 
-    public MCRIIIFImageTargetRotation(boolean mirrored, int degrees) {
+    public MCRIIIFImageTargetRotation(boolean mirrored, double degrees) {
         this.mirrored = mirrored;
         this.degrees = degrees;
     }
@@ -15,7 +15,7 @@ public class MCRIIIFImageTargetRotation {
         return mirrored;
     }
 
-    public int getDegrees() {
+    public double getDegrees() {
         return degrees;
     }
